requirements.txt
Django==3.1.7
django-rest-framework==0.1.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
